@@ -8,20 +8,20 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
   },
   "devDependencies": {
     "@discord-clone/config": "workspace:*",
     "@discord-clone/react": "workspace:*",
     "@discord-clone/tsconfig": "workspace:*",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@vitejs/plugin-react": "^1.0.7",
+    "@types/react": "^18.0.1",
+    "@types/react-dom": "^18.0.0",
+    "@vitejs/plugin-react": "^1.3.0",
     "autoprefixer": "^10.4.4",
     "postcss": "^8.4.12",
     "tailwindcss": "^3.0.23",
-    "typescript": "^4.5.4",
-    "vite": "^2.9.0"
+    "typescript": "^4.6.3",
+    "vite": "^2.9.1"
   }
 }
