@@ -9,8 +9,8 @@
 		"@discord-clone/tsconfig": "workspace:*"
 	},
 	"dependencies": {
-		"@types/react": "^17.0.43",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2"
+		"@types/react": "^18.0.1",
+		"react": "^18.0.0",
+		"react-dom": "^18.0.0"
 	}
 }
