@@ -1,11 +1,9 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import { App } from './App'
 import './index.css'
 
-ReactDOM.render(
-	<React.StrictMode>
-		<App />
-	</React.StrictMode>,
-	document.getElementById('root'),
-)
+const container = document.getElementById('root')
+//@ts-nocheck
+//@ts-ignore
+const root = createRoot(container)
+root.render(<App></App>)
